Revert "ovl: Always reevaluate the file signature for IMA"
This reverts commit 3551d966cc110ea339e6aa89c66e453913484e2d, also known
as:
    [ Upstream commit 18b44bc5a67275641fb26f2c54ba7eef80ac5950 ]

Why: the above commit dramatically increases file access overhead, and
leads to unacceptable performance loss in many applications, for
example, GROMACS. The following GROMACS trivial case runs 2x slower, on
one of our clusters:

    time /usr/local/gromacs/avx2_256/bin/gmx --version &> /dev/null

This also reintroduces the security vulnerability. However, given that
one must have root access to the system in order to exploit this, it is
a very low risk.

Signed-off-by: John Hubbard <jhubbard@nvidia.com>